+++ freqtrade/data/converter.py
