+++ freqtrade/tests/strategy/test_default_strategy.py
@@ -3,7 +3,7 @@ import json
 import pytest
 from pandas import DataFrame
 
-from freqtrade.data.convert import parse_ticker_dataframe
+from freqtrade.data.converter import parse_ticker_dataframe
 from freqtrade.strategy.default_strategy import DefaultStrategy
