+++ freqtrade/tests/test_misc.py
@@ -3,7 +3,7 @@
 import datetime
 from unittest.mock import MagicMock
 
-from freqtrade.data.convert import parse_ticker_dataframe
+from freqtrade.data.converter import parse_ticker_dataframe
 from freqtrade.misc import (common_datearray, datesarray_to_datetimearray,
                             file_dump_json, format_ms_time, shorten_date)
 from freqtrade.optimize.__init__ import load_tickerdata_file
